Moving function_signature.py & fixing dependencies
